refactor(vite-dev-server): extract URL builders to urlPaths.ts, rename to *Url
Reverses the earlier inline and pulls the URL helpers back into a
dedicated file (urlPaths.ts), and harmonizes the names so both end in
Url:

  - getSupportFileRelativePath -> getSupportFileRelativeUrl
  - getSpecRelativeUrl unchanged

Vite's server.warmupRequest signature takes a `url: string`, so the
helpers' return values are URLs by Vite's own naming.

Tests split:
  - test/urlPaths.spec.ts holds the 6 URL builder tests
  - test/devServer.spec.ts keeps just the 5 devServer warmup wiring tests
