+++ npm/vite-dev-server/src/devServer.ts
@@ -3,6 +3,7 @@ import semverMajor from 'semver/functions/major.js'
 import type { UserConfig } from 'vite-7'
 import { getVite, Vite_7, Vite_8 } from './getVite.js'
 import { createViteDevServerConfig, isVite8 } from './resolveConfig.js'
+import { getSpecRelativeUrl, getSupportFileRelativeUrl } from './urlPaths.js'
 
 const debug = debugFn('cypress:vite-dev-server:devServer')
 
@@ -17,57 +18,6 @@ export type ViteDevServerConfig = {
 } & {
   framework?: typeof ALL_FRAMEWORKS[number] // Add frameworks here as we implement
   viteConfig?: ConfigHandler // Derived from the user's vite config
-}
-
-/**
- * Build the support file path matching the URL the browser-side client
- * (client/initCypressTests.js) constructs.
- */
-export function getSupportFileRelativePath (cypressConfig: Cypress.PluginConfigOptions): string {
-  const { projectRoot, supportFile, devServerPublicPathRoute } = cypressConfig
-
-  if (!supportFile) {
-    return ''
-  }
-
-  let supportRelativeToProjectRoot = supportFile.replace(projectRoot, '')
-
-  if (cypressConfig.platform === 'win32') {
-    const platformProjectRoot = projectRoot.replace(/\//g, '\\')
-
-    supportRelativeToProjectRoot = supportFile.replace(platformProjectRoot, '')
-    supportRelativeToProjectRoot = supportRelativeToProjectRoot.replace(/\\/g, '/')
-  }
-
-  const devServerPublicPathBase = devServerPublicPathRoute === '' ? '.' : devServerPublicPathRoute
-
-  return `${devServerPublicPathBase}${supportRelativeToProjectRoot}`
-}
-
-/**
- * Build the spec URL path for Vite's `@fs/` route.
- *
- * The `@fs/` route is mounted at the server root and bypasses the
- * base-aware request middleware, so the path is intentionally returned
- * without the dev server base prefix — `<base>/@fs/<absolute>` produces a
- * "Failed to load url" pre-transform error in Vite 8, while
- * `/@fs/<absolute>` resolves cleanly.
- */
-export function getSpecRelativeUrl (
-  spec: { absolute: string },
-  cypressConfig: Pick<Cypress.PluginConfigOptions, 'platform'>,
-): string {
-  let absolute = spec.absolute
-
-  if (cypressConfig.platform === 'win32') {
-    absolute = absolute.replace(/\\/g, '/')
-  }
-
-  // Strip leading slash so the @fs/ route receives the path the same way the
-  // client constructs it (see client/initCypressTests.js).
-  const normalizedAbsolute = absolute.replace(/^\//, '')
-
-  return `/@fs/${normalizedAbsolute}`
 }
 
 export async function devServer (config: ViteDevServerConfig): Promise<Cypress.ResolvedDevServerConfig> {
@@ -108,10 +58,10 @@ export async function devServer (config: ViteDevServerConfig): Promise<Cypress.R
   // scanner doesn't see, so the optimizer would otherwise first discover
   // them when the browser fetches the spec.
   const warmupTargets: string[] = []
-  const supportPath = getSupportFileRelativePath(config.cypressConfig)
+  const supportFileUrl = getSupportFileRelativeUrl(config.cypressConfig)
 
-  if (supportPath) {
-    warmupTargets.push(supportPath)
+  if (supportFileUrl) {
+    warmupTargets.push(supportFileUrl)
   }
 
   for (const spec of config.specs ?? []) {
